Move DateCell to own file
