+++ 3DCalendar/DateCell.swift
@@ -1,0 +1,56 @@
+//
+//  DateCell.swift
+//  3DCalendar
+//
+//  Created by Najia Haider on 7/2/19.
+//  Copyright © 2019 Najia Haider. All rights reserved.
+//
+
+import UIKit
+
+class DateCell: UICollectionViewCell {
+    
+    let label: UILabel = {
+        let lbl = UILabel()
+        lbl.text = "00"
+        lbl.textAlignment = .center
+        lbl.font = UIFont.systemFont(ofSize: 12)
+        lbl.textColor = UIColor.gray
+        lbl.translatesAutoresizingMaskIntoConstraints = false
+        return lbl
+    }()
+    
+    override init(frame: CGRect) {
+        let frame2 = CGRect(x: 0, y: 0, width: 25, height: 25)
+        super.init(frame: frame2)
+        backgroundColor = UIColor.clear
+        layer.cornerRadius = 5
+        layer.masksToBounds = true
+        
+        setUpViews()
+    }
+    
+    required init?(coder aDecoder: NSCoder) {
+        fatalError("init(coder:) has not been implemented")
+    }
+    
+    func setUpViews() {
+        addSubview(label)
+        label.topAnchor.constraint(equalTo: topAnchor).isActive = true
+        label.leftAnchor.constraint(equalTo: leftAnchor).isActive = true
+        label.rightAnchor.constraint(equalTo: rightAnchor).isActive = true
+        label.bottomAnchor.constraint(equalTo: bottomAnchor).isActive = true
+    }
+}
+
+extension Date {
+    var weekday: Int {
+        return Calendar.current.component(.weekday, from: self)
+    }
+    var firstDayOfTheMonth: Date {
+        return Calendar.current.date(from: Calendar.current.dateComponents([.year,.month], from: self))!
+    }
+    func month(monthIndex: Int) -> String {
+        return Calendar.current.monthSymbols[monthIndex]
+    }
+}
